switch to DNS pointer; Build out Route53 UpsertRR logic

+++ internal/pkg/util/dns/interface.go
@@ -1,5 +1,5 @@
 package dns
 
 type DNS interface {
-  UpsertRR(rrType string, name string, values []string, ttl uint) error
+  UpsertRR(rrType string, name string, values []string, ttl int64) error
 }
